fix compilation onWindows: still installation error
It compiles on Windows, but there is still an issue (Windows only!):
the `ViconDataStreamSDK_CPP.dll` and `ViconDataStreamSDK_CPP.lib` are in the wrong path! They should be under `(venv)/Lib/site-packages/pyvicon_datastream` but they end up in the venv's root folder.
When moving the files manually to the correct location, the module works